refactor: 주석 추가 및 패키지 이동
