+++ src/main/java/roomescape/controller/api/ReservationApiController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.AuthenticatedMember;
-import roomescape.controller.api.validator.IdPositive;
+import roomescape.service.dto.validator.IdPositive;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservationwait.ReservationWaitWithRank;
 import roomescape.domain.member.Member;
